Move _ExtensionEntry to extension.py
